requirements.txt
torch==2.4.0
torchvision==0.19.0
pillow==10.4.0
numpy==1.24.4
scikit-image==0.21.0
transformers==4.43.2
fastapi==0.112.1
uvicorn==0.30.6
starlette==0.38.2
python-multipart==0.0.9
httpx==0.27.0
